+++ Examples/Basic Example/iCarouselExampleViewController.m
@@ -1,0 +1,123 @@
+//
+//  iCarouselExampleViewController.m
+//  iCarouselExample
+//
+//  Created by Nick Lockwood on 03/04/2011.
+//  Copyright 2011 Charcoal Design. All rights reserved.
+//
+
+#import "iCarouselExampleViewController.h"
+
+
+@interface iCarouselExampleViewController ()
+
+@property (nonatomic, retain) NSMutableArray *items;
+
+@end
+
+
+@implementation iCarouselExampleViewController
+
+@synthesize carousel;
+@synthesize items;
+
+- (void)awakeFromNib
+{
+    //set up data
+    //your carousel should always be driven by an array of
+    //data of some kind - don't store data in your item views
+    //or the recycling mechanism will destroy your data once
+    //your item views move off-screen
+    self.items = [NSMutableArray array];
+    for (int i = 0; i < 100; i++)
+    {
+        [items addObject:[NSNumber numberWithInt:i]];
+    }
+}
+
+- (void)dealloc
+{
+    //it's a good idea to set these to nil here to avoid
+    //sending messages to a deallocated viewcontroller
+    carousel.delegate = nil;
+    carousel.dataSource = nil;
+    
+    [carousel release];
+    [items release];
+    [super dealloc];
+}
+
+#pragma mark -
+#pragma mark View lifecycle
+
+- (void)viewDidLoad
+{
+    [super viewDidLoad];
+    
+    //configure carousel
+    carousel.type = iCarouselTypeCoverFlow2;
+}
+
+- (void)viewDidUnload
+{
+    [super viewDidUnload];
+    
+    //free up memory by releasing subviews
+    self.carousel = nil;
+}
+
+- (BOOL)shouldAutorotateToInterfaceOrientation:(UIInterfaceOrientation)interfaceOrientation
+{
+    return YES;
+}
+
+#pragma mark -
+#pragma mark iCarousel methods
+
+- (NSUInteger)numberOfItemsInCarousel:(iCarousel *)carousel
+{
+    //return the total number of items in the carousel
+    return [items count];
+}
+
+- (NSUInteger)numberOfVisibleItemsInCarousel:(iCarousel *)carousel
+{
+    //return the number of visible carousel items on screen
+    //this also affects the appearance of circular-type carousels
+    //this value should be <= numberOfItemsInCarousel
+    //if you have fewer than about 25 items in your carousel, you don't need
+    //to use this method at all (by default it matches numberOfItemsInCarousel)
+    return 25;
+}
+
+- (UIView *)carousel:(iCarousel *)carousel viewForItemAtIndex:(NSUInteger)index reusingView:(UIView *)view
+{
+    UILabel *label = nil;
+    
+    //create new view if no view is available for recycling
+    if (view == nil)
+    {
+        view = [[[UIImageView alloc] initWithImage:[UIImage imageNamed:@"page.png"]] autorelease];
+        view.layer.doubleSided = NO; //prevent back side of view from showing
+        label = [[[UILabel alloc] initWithFrame:view.bounds] autorelease];
+        label.backgroundColor = [UIColor clearColor];
+        label.textAlignment = UITextAlignmentCenter;
+        label.font = [label.font fontWithSize:50];
+        [view addSubview:label];
+    }
+    else
+    {
+        label = [[view subviews] lastObject];
+    }
+    
+    //set item label
+    //remember to always set any properties of your carousel item
+    //views outside of the `if (view == nil) {...}` check otherwise
+    //you'll get weird issues with carousel item content appearing
+    //in the wrong place in the carousel
+    label.text = [[items objectAtIndex:index] stringValue];
+    
+    return view;
+}
+
+@end