+++ Examples/Multiple Carousels/iCarouselExampleViewController.h
@@ -1,0 +1,18 @@
+//
+//  iCarouselExampleViewController.h
+//  iCarouselExample
+//
+//  Created by Nick Lockwood on 03/04/2011.
+//  Copyright 2011 Charcoal Design. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+#import "iCarousel.h"
+
+
+@interface iCarouselExampleViewController : UIViewController <iCarouselDataSource, iCarouselDelegate>
+
+@property (nonatomic, retain) IBOutlet iCarousel *carousel1;
+@property (nonatomic, retain) IBOutlet iCarousel *carousel2;
+
+@end